Complete rewrite that works without rewriting
This approach simply creates templates for each of the fields in the
object or tuple that was passed and gives Nim the task of expanding them
to the right thing. This means that templates called from within a with
section can also use the fields of the object directly, and can fix some
bugs where ambiguity of the symbols would make things hard to rewrite
the body correctly.